requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic[email]==2.5.0
email-validator==2.1.0

# Web Scraping
playwright==1.40.0
selenium==4.15.2

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Database
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2

# Date/Time
python-dateutil==2.8.2

# GitHub Integration
PyGithub==1.59.1

# Authentication & Security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# Payments
stripe==7.5.0

